feat: share authenticaton pages
