+++ apps/partners/src/app/components/sign-out-btn.tsx
