+++ packages/shared/src/lib/auth.ts
