+++ packages/ui/tailwind.config.ts
@@ -8,6 +8,7 @@ const config = {
     "./app/**/*.{ts,tsx}",
     "./src/**/*.{ts,tsx}",
     "../../packages/ui/src/**/*.{ts,tsx}",
+    "../../packages/shared/src/**/*.{ts,tsx}",
   ],
   prefix: "",
   theme: {
